feat(medallion): complete silver 12 models, gold SCD2, warehouse and OLTP fixes plus changelog 0.2.0
- silver: 12 incremental models with dedup, full schema.yml 112 tests, shared incremental_utils macro and generic tests
- gold: SCD2 dims via snapshots, Type1 dims from silver refs, expanded dim_date and fact models, schema rewrites
- bronze: _loaded_date generated columns, reconciled source columns, watermark MERGE fix
- oltp: corrected int/BIGINT types for unit/trailer/distance/weight and added drivers index
- mart: SCD2 driver/vehicle with sk is_current, expanded dims and facts to mirror gold
- orchestration: dbt snapshot before build in DAG and bash/ps scripts
- scripts: env-aware seed URL, powershell strict mode and uv fallback parity, databricks_init fallback

+++ sql/oltp_schema/01_customers.sql
@@ -7,7 +7,7 @@ CREATE TABLE IF NOT EXISTS customers (
     primary_freight_type VARCHAR(100),
     account_status VARCHAR(20),
     contract_start_date DATE,
-    annual_revenue_potential DOUBLE PRECISION,
+    annual_revenue_potential BIGINT,
     updated_at TIMESTAMPTZ NOT NULL DEFAULT now()
 );
 CREATE INDEX IF NOT EXISTS idx_customers_updated_at ON customers (updated_at);
